[Fix/#35] Refactor: Search 컴포넌트의 검색 로직을 useSearchDialog 훅으로 분리
