+++ components/Search/index.tsx
@@ -1,131 +1,37 @@
 "use client";
 
-import { useState, useEffect, useRef, useCallback } from "react";
-import { useRouter } from "next/navigation";
 import Image from "next/image";
 import Icon from "../Icon";
 import Button from "../Button";
 import Post from "@/types/post";
-import { useSearch } from "@/contexts/SearchContext";
+import useSearchDialog from "@/hooks/search";
 
 export default function Search({
   allPosts
 }: {
   allPosts: Post[]
 }) {
-  const router = useRouter();
-
-  const dialogRef = useRef<HTMLDialogElement>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const itemRefs = useRef<(HTMLButtonElement | null)[]>([]);
-  const isKeyboardNav = useRef(false);
-
-  const { isOpen, closeSearch } = useSearch();
-  const [searchKeyword, setSearchKeyword] = useState("");
-  const [selectedIndex, setSelectedIndex] = useState(0);
-  const [displayCount, setDisplayCount] = useState(6);
-
-  const filteredPosts = allPosts.filter(
-    (post) =>
-      post.title.toLowerCase().includes(searchKeyword.toLowerCase()) ||
-      post.content.toLowerCase().includes(searchKeyword.toLowerCase())
-  );
-
-  const displayedPosts = filteredPosts.slice(0, displayCount);
-  const hasMore = filteredPosts.length > displayCount;
-
-  // dialog 열기/닫기 관리
-  useEffect(() => {
-    const dialog = dialogRef.current;
-    if (!dialog) return;
-
-    if (isOpen) {
-      dialog.showModal();
-      inputRef.current?.focus();
-    } else {
-      dialog.close();
-    }
-  }, [isOpen]);
-
-  const handleClose = useCallback(() => {
-    closeSearch();
-    setSearchKeyword("");
-    setSelectedIndex(0);
-    setDisplayCount(6);
-  }, [closeSearch]);
+  const {
+    dialogRef,
+    inputRef,
+    itemRefs,
+    searchKeyword,
+    setSearchKeyword,
+    selectedIndex,
+    displayedPosts,
+    hasMore,
+    handleClose,
+    handleLoadMore,
+    handleMouseEnterItem,
+    handleMouseMove,
+    navigateToPost,
+  } = useSearchDialog(allPosts);
 
   // dialog 닫힐 때 처리
   const handleCancel = (e: React.FormEvent<HTMLDialogElement>) => {
     e.preventDefault();
     handleClose();
   }
-
-  // 배경 클릭 시 닫기
-  useEffect(() => {
-    const dialog = dialogRef.current;
-    if (!dialog) return;
-
-    const handleClickOutside = (e: MouseEvent) => {
-      if (e.target === dialog) {
-        handleClose();
-      }
-    };
-
-    dialog.addEventListener("click", handleClickOutside);
-    return () => dialog.removeEventListener("click", handleClickOutside);
-  }, [handleClose]);
-
-  // 키보드 네비게이션
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (!isOpen) return;
-
-      switch (e.key) {
-        case "Escape":
-          e.preventDefault();
-          handleClose();
-          break;
-        case "ArrowDown":
-          e.preventDefault();
-          isKeyboardNav.current = true;
-          setSelectedIndex((prev) =>
-            prev < displayedPosts.length - 1 ? prev + 1 : prev
-          );
-          break;
-        case "ArrowUp":
-          e.preventDefault();
-          isKeyboardNav.current = true;
-          setSelectedIndex((prev) => (prev > 0 ? prev - 1 : 0));
-          break;
-        case "Enter":
-          e.preventDefault();
-          if (displayedPosts[selectedIndex]) {
-            router.push(`/${displayedPosts[selectedIndex].slug}`);
-            handleClose();
-          }
-          break;
-      }
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [isOpen, displayedPosts, selectedIndex, router, handleClose]);
-
-  // 선택된 항목이 화면에 보이도록 스크롤
-  useEffect(() => {
-    itemRefs.current[selectedIndex]?.scrollIntoView({
-      block: "nearest",
-    });
-  }, [selectedIndex]);
-
-  useEffect(() => {
-    setSelectedIndex(0);
-    setDisplayCount(6);
-  }, [searchKeyword]);
-
-  const handleLoadMore = () => {
-    setDisplayCount((prev) => prev + 6);
-  };
 
   return (
     <dialog
@@ -135,7 +41,7 @@ export default function Search({
         backdrop:bg-background/80
         transition-colors duration-300"
       onCancel={handleCancel}
-      onMouseMove={() => { isKeyboardNav.current = false; }}
+      onMouseMove={handleMouseMove}
     >
       <div className="w-[95vw] lg:w-[90vw] max-w-160 fixed
         top-1/10 lg:top-1/5 left-1/2 -translate-x-1/2
@@ -179,14 +85,8 @@ export default function Search({
                     className={`w-full flex items-center gap-4 px-4 py-3
                       border-none text-left cursor-pointer hover:bg-foreground/5
                       ${index === selectedIndex ? "bg-foreground/5" : ""}`}
-                    onClick={() => {
-                      router.push(`/${post.slug}`);
-                      handleClose();
-                    }}
-                    onMouseEnter={() => {
-                      if (isKeyboardNav.current) return;
-                      setSelectedIndex(index);
-                    }}
+                    onClick={() => navigateToPost(post.slug)}
+                    onMouseEnter={() => handleMouseEnterItem(index)}
                   >
                     <Image
                       src={post.teaser}
